Create htm.code
A malicious user who works with the Dark Web use a certain method to trick unsuspected victims. he started by using three different tags by default. The first tag makes reference to uniformity of links where we have multiple links ranging from 0 to 9 sent to one Email Address , other tags are focused on the Phone Number and Text file on the system. In the beginning for God so love the world................